Add basic variables necessary for non-lin analysis

+++ main_project/component/component.hpp
@@ -30,9 +30,9 @@ public:
 
     // these are used for non linear analysis
     // get the IV characteristics of a component given two nodes
-    virtual float ivAtNode(int n1, int n2, int v) const =0;
+    virtual float ivAtNode(int n1, int n2, float v) const =0;
     // get the derivative of the IV charateristic of a component given two nodes and the node the derivative is based on
-    virtual float divAtNode(int n1, int n2, int dn, int v) const =0;
+    virtual float divAtNode(int n1, int n2, int dn, float v) const =0;
 
     // this should be used to update the value of the voltage and current accross a component after an iteration
     virtual void updateVals(float newVoltage, float newCurrent, int order);
